% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{ma}
\alias{ma}
\title{A Moving Average Filter Function}
\usage{
ma(alpha, data, windowed = F)
}
\arguments{
\item{alpha}{a number that denotes the one-sided window length of the filter.}

\item{data}{a numeric vector to apply the filter}

\item{windowed}{a logical that defaults to FALSE. If true, then it will apply correction for bias caused by seasonal behaviors.}
}
\value{
A numeric vector containing the filtered data
}
\description{
This function computes the centered moving average for filtering.
}
\examples{
ma(2, mydata)
ma(6, mydata, T)
}
